# Codex/master-hive-full.openapi.yaml
openapi: 3.1.0
info:
  title: Master Hive Full Actions
  version: 1.0.0
  description: Full ChatGPT Actions for the Master Hive live server. Includes read and write operations.
servers:
  - url: https://YOUR_PUBLIC_BASE_URL
    description: Replace with PUBLIC_BASE_URL from the live server environment.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    HiveFlowHeader:
      name: X-Hive-Flow
      in: header
      required: true
      description: Identifies ChatGPT Actions traffic in Hive logs.
      schema:
        type: string
        enum:
          - chatgpt
        default: chatgpt
...
security:
  - bearerAuth: []
paths:
...
  /api/oauth-state:
    get:
      operationId: getOAuthState
      summary: Get registered OAuth clients and refresh-token account summary.
      parameters:
        - $ref: "#/components/parameters/HiveFlowHeader"
      responses:
        "200":
          description: OAuth state summary. Does not include token secret values.
          content:
            application/json:
              schema:
                type: object
                properties:
                  clients:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        redirectUris:
                          type: array
                          items:
                            type: string
                      required:
                        - id
                        - redirectUris
                  refreshTokens:
                    type: array
                    items:
                      type: object
                      properties:
                        email:
                          type: string
                      required:
                        - email
                required:
                  - clients
                  - refreshTokens